% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{AIC.qlfit}
\alias{AIC.qlfit}
\title{AIC for qlfit object}
\usage{
\method{AIC}{qlfit}(object, ..., k = 2)
}
\arguments{
\item{object}{A qlfit object}

\item{...}{Additional arguments (ignored)}

\item{k}{Penalty parameter (default 2)}
}
\value{
AIC value
}
\description{
AIC for qlfit object
}
