% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BIC.qlfit}
\alias{BIC.qlfit}
\title{BIC for qlfit object}
\usage{
\method{BIC}{qlfit}(object, ...)
}
\arguments{
\item{object}{A qlfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
BIC value
}
\description{
BIC for qlfit object
}
