% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_probs}
\alias{bound_probs}
\title{Bound probabilities away from 0 and 1}
\usage{
bound_probs(p, eps = 1e-10)
}
\arguments{
\item{p}{Probability vector or matrix}

\item{eps}{Minimum distance from 0 and 1 (default 1e-10)}
}
\value{
Bounded probabilities
}
\description{
Bound probabilities away from 0 and 1
}
\keyword{internal}
