% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{build_class_monotonicity_fn}
\alias{build_class_monotonicity_fn}
\title{Build class monotonicity constraint function}
\usage{
build_class_monotonicity_fn(n_items, n_classes)
}
\arguments{
\item{n_items}{Number of items}

\item{n_classes}{Number of latent classes}
}
\value{
Function that takes item_probs matrix and returns constraint values (>= 0 if satisfied)
}
\description{
Constraint: β_ic ≤ β_ic' for all items i and classes c < c'
This means item probabilities increase across classes.
}
\keyword{internal}
