% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{build_constraint_fn}
\alias{build_constraint_fn}
\title{Build constraint function from specification}
\usage{
build_constraint_fn(constraints, n_items, n_classes, item_order = NULL)
}
\arguments{
\item{constraints}{A ql_constraints object}

\item{n_items}{Number of items}

\item{n_classes}{Number of latent classes}

\item{item_order}{Item order (may be estimated if not provided)}
}
\value{
Constraint function
}
\description{
Build constraint function from specification
}
\keyword{internal}
