% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-class.R
\docType{class}
\name{checks-class}
\alias{checks-class}
\alias{checks}
\title{Class "checks"}
\description{
The formal S4 class for checks. This class contains transformed versions
of the raw response data as well as summaries of the checks.
}
\details{
Objects of class \code{checks} contain all information returned by
\code{\link[=ConjointChecks]{ConjointChecks()}} (and by the related wrappers \code{\link[=SingleCancel]{SingleCancel()}},
\code{\link[=DoubleCancel]{DoubleCancel()}}, and \code{\link[=TripleCancel]{TripleCancel()}}).
}
\section{Slots}{

\describe{
\item{\code{N}}{matrix containing the number of respondents at each item/ability
intersection.}

\item{\code{n}}{matrix containing the number of correct responses at each
item/ability intersection.}

\item{\code{Checks}}{list containing information about each checked submatrix.}

\item{\code{tab}}{matrix containing information about the detected violations at
each item/ability intersection.}

\item{\code{means}}{vector containing weighted and unweighted means for the
detected violations (where weights are the number of individuals at each
ability level).}

\item{\code{check.counts}}{matrix giving the number of times an item/ability cell
was sampled.}
}}

\seealso{
\code{\link[=ConjointChecks]{ConjointChecks()}}, \code{\link[=summary.checks]{summary.checks()}}, \code{\link[=plot.checks]{plot.checks()}}
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
\keyword{classes}
