% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{class_assignments}
\alias{class_assignments}
\title{Get modal class assignments}
\usage{
class_assignments(object)
}
\arguments{
\item{object}{A qlfit object}
}
\value{
Vector of class assignments (1 to n_classes)
}
\description{
Get modal class assignments
}
