% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{classification_accuracy}
\alias{classification_accuracy}
\title{Compute classification accuracy}
\usage{
classification_accuracy(object)
}
\arguments{
\item{object}{A qlfit object}
}
\value{
Mean posterior probability for modal class assignments
}
\description{
Modal assignment accuracy based on posterior probabilities
}
