% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{compare_fit}
\alias{compare_fit}
\title{Compare fit of multiple models}
\usage{
compare_fit(..., measures = "ic")
}
\arguments{
\item{...}{qlfit objects to compare}

\item{measures}{Which measures to include: "all", "ic" (information criteria only),
or a character vector of specific measures}
}
\value{
Data frame with fit measures for all models
}
\description{
Compare fit of multiple models
}
