% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{compute_expected_frequencies}
\alias{compute_expected_frequencies}
\title{Compute expected pattern frequencies}
\usage{
compute_expected_frequencies(item_probs, class_probs, patterns)
}
\arguments{
\item{item_probs}{Item probability matrix (I x C)}

\item{class_probs}{Class probability vector}

\item{patterns}{Character vector of response patterns}
}
\value{
Vector of expected frequencies
}
\description{
Compute expected pattern frequencies
}
\keyword{internal}
