% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{compute_loglik}
\alias{compute_loglik}
\title{Compute observed data log-likelihood}
\usage{
compute_loglik(data, item_probs, class_probs)
}
\arguments{
\item{data}{Binary data matrix}

\item{item_probs}{Item probability matrix (I x C)}

\item{class_probs}{Class probability vector}
}
\value{
Log-likelihood value
}
\description{
Compute observed data log-likelihood
}
\keyword{internal}
