% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{compute_rasch_probs}
\alias{compute_rasch_probs}
\title{Compute item probabilities from Rasch parameters}
\usage{
compute_rasch_probs(theta, delta)
}
\arguments{
\item{theta}{Class locations (vector of length C)}

\item{delta}{Item difficulties (vector of length I)}
}
\value{
Item probability matrix (I x C)
}
\description{
Compute item probabilities from Rasch parameters
}
\keyword{internal}
