% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{count_constraints}
\alias{count_constraints}
\title{Count number of constraints}
\usage{
count_constraints(constraints, n_items, n_classes)
}
\arguments{
\item{constraints}{A ql_constraints object}

\item{n_items}{Number of items}

\item{n_classes}{Number of latent classes}
}
\value{
Number of inequality constraints
}
\description{
Count number of constraints
}
\keyword{internal}
