% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_parameters}
\alias{count_parameters}
\title{Count number of free parameters}
\usage{
count_parameters(model_type, n_items, n_classes)
}
\arguments{
\item{model_type}{Model type code}

\item{n_items}{Number of items}

\item{n_classes}{Number of classes}
}
\value{
Number of free parameters
}
\description{
Count number of free parameters
}
\keyword{internal}
