% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{e_step}
\alias{e_step}
\title{E-step: Compute posterior class memberships}
\usage{
e_step(data, item_probs, class_probs)
}
\arguments{
\item{data}{Binary data matrix (n x I)}

\item{item_probs}{Item probability matrix (I x C)}

\item{class_probs}{Class probability vector (length C)}
}
\value{
List with posteriors and log-likelihood
}
\description{
E-step: Compute posterior class memberships
}
\keyword{internal}
