% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_item_order}
\alias{estimate_item_order}
\title{Estimate item ordering from data}
\usage{
estimate_item_order(data)
}
\arguments{
\item{data}{Binary data matrix}
}
\value{
Vector of item indices ordered from easiest to hardest
}
\description{
Estimate item ordering from data
}
\keyword{internal}
