% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dm.R
\name{fit_dm_projection}
\alias{fit_dm_projection}
\title{Fit DM model using projection method}
\usage{
fit_dm_projection(
  data,
  n_classes,
  item_order = NULL,
  n_starts = 10,
  max_iter = 500,
  tol = 1e-06,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of classes}

\item{item_order}{Item order}

\item{n_starts}{Number of random starts}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{seed}{Random seed}

\item{verbose}{Print progress}
}
\value{
qlfit object
}
\description{
Fit DM model using projection method
}
\keyword{internal}
