% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcr.R
\name{fit_lcr_mle}
\alias{fit_lcr_mle}
\title{Fit LCR model using joint MLE}
\usage{
fit_lcr_mle(data, n_classes, n_starts = 10, seed = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of classes}

\item{n_starts}{Number of random starts}

\item{seed}{Random seed}

\item{verbose}{Print progress}
}
\value{
qlfit object
}
\description{
Alternative fitting approach using direct likelihood optimization
rather than EM.
}
\keyword{internal}
