% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{fit_measures}
\alias{fit_measures}
\title{Fit Measures for QuantFit Models}
\usage{
fit_measures(object, ...)
}
\arguments{
\item{object}{A qlfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
Data frame with fit measures
}
\description{
Functions for computing model fit statistics including
log-likelihood, AIC, BIC, SABIC, and related measures
}
\examples{
\dontrun{
fit <- fit_un(data, n_classes = 3)
fit_measures(fit)
}

}
