% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_un.R
\name{fit_un_single}
\alias{fit_un_single}
\title{Fit UN model with specific starting values}
\usage{
fit_un_single(
  data,
  n_classes,
  init_probs,
  init_class_probs,
  max_iter = 1000,
  tol = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of classes}

\item{init_probs}{Starting item probabilities}

\item{init_class_probs}{Starting class probabilities}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
qlfit object
}
\description{
Fit UN model with specific starting values
}
\keyword{internal}
