% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_model}
\alias{get_model}
\title{Extract a fitted model from comparison}
\usage{
get_model(object, model)
}
\arguments{
\item{object}{A qlcompare object}

\item{model}{Model name to extract}
}
\value{
A qlfit object
}
\description{
Extract a fitted model from comparison
}
