% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_bootstrap_null.R
\name{\%||\%}
\alias{\%||\%}
\title{Print method for ccnull objects}
\usage{
a \%||\% b
}
\arguments{
\item{x}{A ccnull object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for ccnull objects
}
