% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_item_probs}
\alias{init_item_probs}
\title{Initialize item probabilities for LCA}
\usage{
init_item_probs(
  data,
  n_classes,
  method = c("random", "kmeans", "quantiles"),
  seed = NULL
)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of latent classes}

\item{method}{Initialization method: "random", "kmeans", or "quantiles"}

\item{seed}{Random seed (optional)}
}
\value{
Matrix of item probabilities (n_items x n_classes)
}
\description{
Initialize item probabilities for LCA
}
\keyword{internal}
