% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dm.R
\name{init_item_probs_dm}
\alias{init_item_probs_dm}
\title{Initialize item probabilities satisfying double monotonicity}
\usage{
init_item_probs_dm(data, n_classes, item_order, seed = NULL)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of classes}

\item{item_order}{Item order (easiest to hardest)}

\item{seed}{Random seed}
}
\value{
Item probability matrix satisfying both DM constraints
}
\description{
Initialize item probabilities satisfying double monotonicity
}
\keyword{internal}
