% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_iio.R
\name{init_item_probs_iio}
\alias{init_item_probs_iio}
\title{Initialize item probabilities satisfying IIO constraints}
\usage{
init_item_probs_iio(data, n_classes, item_order, seed = NULL)
}
\arguments{
\item{data}{Data matrix}

\item{n_classes}{Number of classes}

\item{item_order}{Item order (easiest to hardest)}

\item{seed}{Random seed}
}
\value{
Item probability matrix satisfying IIO constraints
}
\description{
Initialize item probabilities satisfying IIO constraints
}
\keyword{internal}
