% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{interpret_structure}
\alias{interpret_structure}
\title{Interpret latent structure based on best model}
\usage{
interpret_structure(model)
}
\arguments{
\item{model}{Model code (UN, MON, IIO, DM, LCR, RM)}
}
\value{
Character string with interpretation
}
\description{
Interpret latent structure based on best model
}
\keyword{internal}
