% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit (logistic) function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector of probabilities
}
\description{
Inverse logit (logistic) function
}
\keyword{internal}
