% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcr.R
\name{lcr_scores}
\alias{lcr_scores}
\title{Extract Rasch-scale scores from LCR model}
\usage{
lcr_scores(object, type = c("eap", "modal"))
}
\arguments{
\item{object}{A qlfit object from fit_lcr}

\item{type}{Type of score: "eap" (expected a posteriori) or "modal"}
}
\value{
Vector of estimated person scores on the theta scale
}
\description{
Extract Rasch-scale scores from LCR model
}
