% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks-class.R
\docType{class}
\name{list.null-class}
\alias{list.null-class}
\alias{list.null}
\title{Class "list.null"}
\description{
The formal S4 class union for \code{list.null}. This class contains either a
list or \code{NULL} and is used internally to hold the results of the
individual cancellation checks.
}
\details{
Objects of this class are used internally.
}
\seealso{
\code{\link[=ConjointChecks]{ConjointChecks()}}, \code{\link[=summary.checks]{summary.checks()}}, \code{\link[=plot.checks]{plot.checks()}}
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
\keyword{classes}
