% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Compute log-sum-exp in a numerically stable way}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
log(sum(exp(x)))
}
\description{
Compute log-sum-exp in a numerically stable way
}
\keyword{internal}
