% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{logLik.qlfit}
\alias{logLik.qlfit}
\title{Log-likelihood for qlfit object}
\usage{
\method{logLik}{qlfit}(object, ...)
}
\arguments{
\item{object}{A qlfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
Log-likelihood value with attributes
}
\description{
Log-likelihood for qlfit object
}
