% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{m_step}
\alias{m_step}
\title{M-step: Update parameters (unconstrained)}
\usage{
m_step(data, posteriors)
}
\arguments{
\item{data}{Binary data matrix (n x I)}

\item{posteriors}{Posterior class membership matrix (n x C)}
}
\value{
List with updated item_probs and class_probs
}
\description{
M-step: Update parameters (unconstrained)
}
\keyword{internal}
