% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiple_starts}
\alias{multiple_starts}
\title{Multiple random starts wrapper}
\usage{
multiple_starts(fit_fn, n_starts = 10, seed = NULL, verbose = FALSE, ...)
}
\arguments{
\item{fit_fn}{Fitting function to call}

\item{n_starts}{Number of random starts}

\item{seed}{Base random seed}

\item{verbose}{Print progress}

\item{...}{Arguments passed to fit_fn}
}
\value{
Best fit across all starts
}
\description{
Multiple random starts wrapper
}
\keyword{internal}
