% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_qlcompare}
\alias{new_qlcompare}
\title{Create a qlcompare object}
\usage{
new_qlcompare(fits, comparison_table, best_model, call = NULL)
}
\arguments{
\item{fits}{List of qlfit objects}

\item{comparison_table}{Data frame with comparison statistics}

\item{best_model}{Name of best model by BIC}

\item{call}{Original function call}
}
\value{
Object of class "qlcompare"
}
\description{
Create a qlcompare object
}
\keyword{internal}
