% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_qlfit}
\alias{new_qlfit}
\title{Create a qlfit object}
\usage{
new_qlfit(
  model_type,
  item_probs,
  class_probs,
  posteriors,
  loglik,
  n_par,
  n_obs,
  n_items,
  n_classes,
  convergence,
  iterations,
  call = NULL,
  theta = NULL,
  delta = NULL,
  item_order = NULL,
  constraints = NULL,
  se = NULL
)
}
\arguments{
\item{model_type}{Character string indicating model type
("UN", "MON", "IIO", "DM", "LCR", "RM")}

\item{item_probs}{Matrix of item probabilities (items x classes)}

\item{class_probs}{Vector of class probabilities}

\item{posteriors}{Matrix of posterior class memberships (n x classes)}

\item{loglik}{Log-likelihood value}

\item{n_par}{Number of estimated parameters}

\item{n_obs}{Number of observations}

\item{n_items}{Number of items}

\item{n_classes}{Number of latent classes}

\item{convergence}{Logical indicating convergence}

\item{iterations}{Number of EM iterations}

\item{call}{Original function call}

\item{theta}{Class locations (for LCR/RM models)}

\item{delta}{Item difficulties (for LCR/RM models)}

\item{item_order}{Item ordering (for IIO/DM models)}

\item{constraints}{List of active constraints}

\item{se}{Standard errors (if computed)}
}
\value{
Object of class "qlfit"
}
\description{
Create a qlfit object
}
\keyword{internal}
