% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TripleCancel.R
\name{omni.check_triple}
\alias{omni.check_triple}
\title{Gibbs sampler for triple cancellation in one 4x4 submatrix}
\usage{
omni.check_triple(N, n, n.iter, burn = 1000, CR)
}
\arguments{
\item{N}{Matrix containing the total number of responses.}

\item{n}{Matrix containing the number of correct responses.}

\item{n.iter}{Total number of samples.}

\item{burn}{Number of initial samples that should be discarded.}

\item{CR}{Width of the credible region taken from the posterior, e.g. a
95\% credible region is \code{c(.025,.975)}.}
}
\value{
A list with elements \code{low}, \code{high}, and \code{mean}: matrices giving the
lower and upper limits of the credible region and the posterior mean for
each cell.
}
\description{
Internal function. Checks ONLY the triple cancellation bits in one 4x4
submatrix, using the 14 coherent triple cancellation tests of Kyngdon and
Richards (2006).
}
\references{
Perline, R., Wright, B. D., & Wainer, H. (1979). The Rasch model as
additive conjoint measurement. \emph{Applied Psychological Measurement},
3(2), 237-255.
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
\keyword{internal}
