% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern_frequencies}
\alias{pattern_frequencies}
\title{Compute response pattern frequencies}
\usage{
pattern_frequencies(data)
}
\arguments{
\item{data}{Binary data matrix}
}
\value{
Data frame with patterns and frequencies
}
\description{
Compute response pattern frequencies
}
\keyword{internal}
