% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern_to_vector}
\alias{pattern_to_vector}
\title{Convert pattern string back to numeric vector}
\usage{
pattern_to_vector(pattern)
}
\arguments{
\item{pattern}{Character string of 0s and 1s}
}
\value{
Numeric vector
}
\description{
Convert pattern string back to numeric vector
}
\keyword{internal}
