% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{pearson_chisq}
\alias{pearson_chisq}
\title{Compute Pearson chi-square}
\usage{
pearson_chisq(object, data)
}
\arguments{
\item{object}{A qlfit object}

\item{data}{Original data matrix}
}
\value{
Chi-square value with degrees of freedom
}
\description{
X² = sum((O - E)² / E)
}
