% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_bootstrap_null.R
\name{plot.ccnull}
\alias{plot.ccnull}
\title{Plot method for ccnull objects}
\usage{
\method{plot}{ccnull}(x, ...)
}
\arguments{
\item{x}{A ccnull object.}

\item{...}{Passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
Invisibly returns x.
}
\description{
Histogram of the Rasch null distribution with the observed rate marked.
}
