% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.checks}
\alias{plot.checks}
\title{Plot checks produced by ConjointChecks}
\usage{
\method{plot}{checks}(x, items = NULL, item.labels = TRUE, ...)
}
\arguments{
\item{x}{Object returned by \code{\link[=ConjointChecks]{ConjointChecks()}} of class
\code{\link[=checks-class]{checks}}.}

\item{items}{Vector of item numbers to include in a single plot. Defaults
to all, but this is less helpful for diagnostic purposes.}

\item{item.labels}{Should item numbers be included? Defaults to \code{TRUE}. If
length of \code{items} is unity (perhaps if the small multiple format of
Tufte, 2001 is going to be used), then the item number gets printed
below the x-axis. If the length of \code{items} is more than unity, the item
number gets plotted in the figure above the largest proportion of
violations for each item.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Takes output from \code{\link[=ConjointChecks]{ConjointChecks()}} and produces a
\link[graphics:matplot]{matplot} showing the percentage of reported violations
at each cell.
}
\examples{
\dontrun{
opar <- par()
par(mfrow=c(3,2))
plot(rasch1000)
plot(rasch1000,items=c(5,10,15))
for (i in c(3,9,13,18)) plot(rasch1000,items=i)
par(opar)
}

}
\references{
Tufte, E. R. (2001). The visual display of quantitative information
(2nd ed.). Cheshire, CT: Graphics Press.
}
\seealso{
\code{\link[=ConjointChecks]{ConjointChecks()}}, \code{\link[=summary.checks]{summary.checks()}}
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
