% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.qlcompare}
\alias{plot.qlcompare}
\title{Plot method for qlcompare objects}
\usage{
\method{plot}{qlcompare}(x, ...)
}
\arguments{
\item{x}{A qlcompare object}

\item{...}{Additional arguments passed to plot_comparison}
}
\value{
Invisibly returns x
}
\description{
Plot method for qlcompare objects
}
