% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.qlfit}
\alias{plot.qlfit}
\title{Plot method for qlfit objects}
\usage{
\method{plot}{qlfit}(x, type = c("irf", "class"), ...)
}
\arguments{
\item{x}{A qlfit object}

\item{type}{Type of plot: "irf" (default) or "class"}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
Invisibly returns x
}
\description{
Plot method for qlfit objects
}
