% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_irfs_rm}
\alias{plot_irfs_rm}
\title{Plot IRFs for Rasch Model (continuous theta)}
\usage{
plot_irfs_rm(
  object,
  items = NULL,
  legend_pos = "right",
  colors = NULL,
  main = NULL,
  ...
)
}
\description{
Plot IRFs for Rasch Model (continuous theta)
}
\keyword{internal}
