% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_polytomous.R
\name{poly_estep}
\alias{poly_estep}
\title{Polytomous E-step}
\usage{
poly_estep(data, item_probs, class_probs)
}
\arguments{
\item{data}{integer matrix (n x J), entries 0..m_j}

\item{item_probs}{list length J; \code{item_probs[[j]]} is C x (m_j + 1), rows sum 1}

\item{class_probs}{length C}
}
\description{
Polytomous E-step
}
\keyword{internal}
