% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_bootstrap_null.R
\name{print.cchier}
\alias{print.cchier}
\title{Print method for cchier objects}
\usage{
\method{print}{cchier}(x, ...)
}
\arguments{
\item{x}{A cchier object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns x.
}
\description{
Print method for cchier objects
}
