% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{print.lr_test}
\alias{print.lr_test}
\title{Print method for lr_test}
\usage{
\method{print}{lr_test}(x, ...)
}
\arguments{
\item{x}{An lr_test object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for lr_test
}
