% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qlcompare}
\alias{print.qlcompare}
\title{Print method for qlcompare objects}
\usage{
\method{print}{qlcompare}(x, ...)
}
\arguments{
\item{x}{A qlcompare object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for qlcompare objects
}
