% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ll.R
\name{print.qleqtest}
\alias{print.qleqtest}
\title{Print method for qleqtest objects}
\usage{
\method{print}{qleqtest}(x, ...)
}
\arguments{
\item{x}{A qleqtest object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for qleqtest objects
}
