% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qlfit}
\alias{print.qlfit}
\title{Print method for qlfit objects}
\usage{
\method{print}{qlfit}(x, ...)
}
\arguments{
\item{x}{A qlfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for qlfit objects
}
