% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_n_classes.R
\name{print.qlnclasses}
\alias{print.qlnclasses}
\title{Print method for qlnclasses objects}
\usage{
\method{print}{qlnclasses}(x, ...)
}
\arguments{
\item{x}{A qlnclasses object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for qlnclasses objects
}
