% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{print.qlse}
\alias{print.qlse}
\title{Print method for qlse objects}
\usage{
\method{print}{qlse}(x, digits = 4, ...)
}
\arguments{
\item{x}{A qlse object from \code{\link{compute_se}}}

\item{digits}{Number of digits to display (default 4)}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for qlse objects
}
