% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ll.R
\name{print.qlselect_ll}
\alias{print.qlselect_ll}
\title{Print method for qlselect_ll objects}
\usage{
\method{print}{qlselect_ll}(x, ...)
}
\arguments{
\item{x}{A qlselect_ll object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print method for qlselect_ll objects
}
