% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress_bar}
\alias{progress_bar}
\title{Print progress bar}
\usage{
progress_bar(iteration, total, width = 50)
}
\arguments{
\item{iteration}{Current iteration}

\item{total}{Total iterations}

\item{width}{Width of progress bar}
}
\description{
Print progress bar
}
\keyword{internal}
