% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ll.R
\name{refit_model_type}
\alias{refit_model_type}
\title{Refit a latent class model of a given type}
\usage{
refit_model_type(model_type, data, n_classes, n_starts, use_cpp)
}
\arguments{
\item{model_type}{One of "UN", "MON", "IIO", "DM", "LCR".}

\item{data}{Binary response matrix.}

\item{n_classes}{Number of latent classes.}

\item{n_starts}{Number of random starts.}

\item{use_cpp}{Use the compiled EM engine.}
}
\value{
A qlfit object.
}
\description{
Refit a latent class model of a given type
}
\keyword{internal}
