% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures.R
\name{relative_fit}
\alias{relative_fit}
\title{Compute relative fit indices}
\usage{
relative_fit(object, data)
}
\arguments{
\item{object}{A qlfit object}

\item{data}{Original data matrix}
}
\value{
List with fit indices
}
\description{
Computes CFI, TLI, and RMSEA (approximations for LCA)
}
