% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rm.R
\name{rm_item_info}
\alias{rm_item_info}
\title{Item information function for Rasch model}
\usage{
rm_item_info(object, theta = seq(-4, 4, by = 0.1))
}
\arguments{
\item{object}{A qlfit object from fit_rm}

\item{theta}{Vector of theta values at which to compute information}
}
\value{
Matrix of item information values (items x theta)
}
\description{
Item information function for Rasch model
}
