% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rm.R
\name{rm_scores}
\alias{rm_scores}
\title{Extract Rasch scores from RM model}
\usage{
rm_scores(object, type = c("EAP", "MAP", "ML", "WLE"))
}
\arguments{
\item{object}{A qlfit object from fit_rm}

\item{type}{Type of score: "EAP" (expected a posteriori), "MAP" (maximum a posteriori),
"ML" (maximum likelihood), or "WLE" (weighted likelihood)}
}
\value{
Data frame with person scores and standard errors
}
\description{
Extract Rasch scores from RM model
}
