% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rm.R
\name{rm_test_info}
\alias{rm_test_info}
\title{Test information function for Rasch model}
\usage{
rm_test_info(object, theta = seq(-4, 4, by = 0.1))
}
\arguments{
\item{object}{A qlfit object from fit_rm}

\item{theta}{Vector of theta values}
}
\value{
Vector of test information values
}
\description{
Test information function for Rasch model
}
