% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_log_sum_exp}
\alias{row_log_sum_exp}
\title{Compute log-sum-exp for matrix rows}
\usage{
row_log_sum_exp(x)
}
\arguments{
\item{x}{Numeric matrix}
}
\value{
Vector of log(sum(exp(x))) for each row
}
\description{
Compute log-sum-exp for matrix rows
}
\keyword{internal}
