% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{sabic}
\alias{sabic}
\title{Sample-adjusted BIC (SABIC) for qlfit object}
\usage{
sabic(object)
}
\arguments{
\item{object}{A qlfit object}
}
\value{
SABIC value
}
\description{
Sample-adjusted BIC (SABIC) for qlfit object
}
