% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCancel.R
\name{SingleCancel}
\alias{SingleCancel}
\title{Check single cancellation in all adjacent 3x3 submatrices}
\usage{
SingleCancel(N, n, CR = c(0.025, 0.975), single, mc.cores = 1)
}
\arguments{
\item{N}{Matrix containing the total number of responses.}

\item{n}{Matrix containing the number of correct responses.}

\item{CR}{Width of the credible region taken from the posterior. Defaults
to a 95\% credible region (\code{c(.025,.975)}).}

\item{single}{Direction of the single cancellation constraint: \code{"R"} for
rows or \code{"C"} for columns.}

\item{mc.cores}{The number of cores to parallelize over.}
}
\value{
An object of class \code{\link[=checks-class]{checks}} summarizing the
detected violations.
}
\description{
Checks ONLY the single cancellation axiom of additive conjoint measurement
in all adjacently formed 3x3 submatrices of \code{n}/\code{N}, in a single direction
(across rows or across columns).
}
\examples{
\dontrun{
# simulated Rasch example
n.items <- 20
n.respondents <- 2000
diff <- rnorm(n.items)
abil <- rnorm(n.respondents)
kern <- outer(abil, diff, "-")
pv <- exp(kern)/(1+exp(kern))
resp <- ifelse(pv > runif(n.items*n.respondents), 1, 0)
tmp <- PrepareChecks(resp)
out <- SingleCancel(tmp$N, tmp$n, single = "R")
summary(out)
}

}
\references{
Perline, R., Wright, B. D., & Wainer, H. (1979). The Rasch model as
additive conjoint measurement. \emph{Applied Psychological Measurement},
3(2), 237-255.
}
\seealso{
\code{\link[=ConjointChecks]{ConjointChecks()}}, \code{\link[=DoubleCancel]{DoubleCancel()}}, \code{\link[=TripleCancel]{TripleCancel()}}
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
