% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softmax}
\alias{softmax}
\title{Softmax function}
\usage{
softmax(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector of probabilities summing to 1
}
\description{
Softmax function
}
\keyword{internal}
