% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{standard_errors}
\alias{standard_errors}
\title{Standard Errors for QuantFit Models}
\description{
Observed-information (numerical Hessian) and nonparametric
bootstrap standard errors for fitted QuantFit models.
}
