% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.checks}
\alias{summary.checks}
\title{Summarize checks produced by ConjointChecks}
\usage{
\method{summary}{checks}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link[=ConjointChecks]{ConjointChecks()}} of class
\code{\link[=checks-class]{checks}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with the weighted and unweighted mean violation rates
(\code{Means}) and the mean violation rate per item (\code{items}).
}
\description{
Takes output from \code{\link[=ConjointChecks]{ConjointChecks()}} and produces summary measures of the
reported violations.
}
\examples{
\dontrun{
summary(rasch1000)
}

}
\seealso{
\code{\link[=ConjointChecks]{ConjointChecks()}}, \code{\link[=plot.checks]{plot.checks()}}
}
\author{
Ben Domingue \email{ben.domingue@gmail.com}
}
