% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.qlfit}
\alias{summary.qlfit}
\title{Summary method for qlfit objects}
\usage{
\method{summary}{qlfit}(object, ...)
}
\arguments{
\item{object}{A qlfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns a summary list
}
\description{
Summary method for qlfit objects
}
