% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data}
\alias{validate_data}
\title{Validate input data for latent class models}
\usage{
validate_data(data, allow_na = FALSE)
}
\arguments{
\item{data}{Matrix or data frame of binary responses (0/1)}

\item{allow_na}{Logical, whether to allow missing values}
}
\value{
Validated data matrix
}
\description{
Validate input data for latent class models
}
\keyword{internal}
