% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_se}
\alias{compute_se}
\title{Compute standard errors via numerical Hessian}
\usage{
compute_se(fit, data)
}
\arguments{
\item{fit}{A qlfit object}

\item{data}{The original data}
}
\value{
Standard errors for parameters
}
\description{
Compute standard errors via numerical Hessian
}
\keyword{internal}
