% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{m_step_rasch}
\alias{m_step_rasch}
\title{M-step for Latent Class Rasch model}
\usage{
m_step_rasch(data, posteriors, theta, delta, class_probs)
}
\arguments{
\item{data}{Binary data matrix}

\item{posteriors}{Posterior class memberships}

\item{theta}{Current theta values}

\item{delta}{Current delta values}

\item{class_probs}{Current class probabilities}
}
\value{
List with updated parameters
}
\description{
M-step for Latent Class Rasch model
}
\keyword{internal}
