% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rm.R
\name{rm_itemfit}
\alias{rm_itemfit}
\title{Item fit statistics for Rasch / partial-credit model}
\usage{
rm_itemfit(object)
}
\arguments{
\item{object}{A qlfit object from fit_rm}
}
\value{
Data frame with outfit and infit mean-square statistics per item.
}
\description{
Item fit statistics for Rasch / partial-credit model
}
