% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rm.R
\name{rm_personfit}
\alias{rm_personfit}
\title{Person fit statistics for Rasch model}
\usage{
rm_personfit(object)
}
\arguments{
\item{object}{A qlfit object from fit_rm}
}
\value{
Data frame with person fit statistics
}
\description{
Person fit statistics for Rasch model
}
